refactor: organized the training event into it's own file
