+++ orchestrator/src/lib.rs
@@ -9,8 +9,7 @@ use comms::Connector;
 
 use configs::{Adapter, DataSrc, ModelConfig, TrainingConfig, Validator};
 use dataset_format::{DatasetFormat, convert_to_binary};
-use error::{OrchErr, Result};
-
+pub use error::{OrchErr, Result};
 pub use sessions::{CancelHandle, Session, StopReason, TrainedModel, TrainingEvent};
 
 /// Starts the distributed training process and returns an active session.
