+++ orchestrator/src/sessions/training_strategies/strategy.rs
@@ -1,0 +1,1 @@
+pub trait TrainingStrategy {}
